Refactor: 토너먼트 페이지 리팩토링
코드 리팩토링
- TournamentPage : mount() 할때, 실행되던 function들을 component/TournamentBracket 으로 이동
- component/TournamentBracket : class형식에서 function으로 수정
- TournamentPage : Mock Data / 나중에 받아 올 데이터의 형식과 비슷하게 수정
애니메이션 효과 적용
- 'PONG !' 문자열 데이터를 받으면, pong-animation 클래스 추가 되어 적용 (다음에 대결 해야 하는 대진 테이블)
- winPlayer 데이터를 받고, 문자열 데이터가 winPlayer와 같다면, win-player-animation 클래스 추가 되어 적용 (부전승 or 대결 후 이긴 플레이어 대진 테이블)
