+++ FE/public/css/styles.css
@@ -702,6 +702,9 @@ td {
 }
 
 .user-wrapper-position {
+	display: flex;
+	justify-content: center;
+	align-items: center;
 	width: 100%;
 	height: 4rem;
 	padding: 0 1rem;
@@ -709,6 +712,39 @@ td {
 	border-radius: 1rem;
 	border: 0.2rem solid #ffffff;
 	position: absolute;
+}
+
+.pong-animation {
+	font-weight: bold;
+	animation: pongAnimation 0.5s infinite alternate;
+}
+@keyframes pongAnimation {
+	0% {
+		box-shadow: 0rem 0rem 1rem 0.5rem #ffffff;
+		text-shadow: 0rem 0rem 1rem #ff5d84;
+	}
+	100% {
+		box-shadow: 0rem 0rem 1rem 0.5rem #ff5d84;
+		text-shadow: 0rem 0rem 1rem #ffffff;
+	}
+}
+
+.win-player-animation {
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	font-weight: bold;
+	overflow: hidden;
+	white-space: nowrap;
+	animation: typingAnimation 3s steps(40) forwards;
+}
+@keyframes typingAnimation {
+	from {
+		width: 0;
+	}
+	to {
+		width: 100%;
+	}
 }
 
 .wire-container {
